*: refactor and cleanup
This shouldn't change functionality. It mainly consists of code cleanup
and refactoring.
