+++ src/checks/max_summary_length.rs
@@ -16,7 +16,7 @@ pub struct MaxSummaryLength;
 
 use errors::*;
 use config::Checks;
-use github_handler::CommitData;
+use github::CommitData;
 use checks::Check;
 
 impl Check for MaxSummaryLength {
