+++ src/checks/no_capitalize_summary.rs
@@ -19,7 +19,7 @@ pub struct NoCapitalizeSummary;
 use self::regex::Regex;
 use errors::*;
 use config::Checks;
-use github_handler::CommitData;
+use github::CommitData;
 use checks::Check;
 
 impl Check for NoCapitalizeSummary {
