+++ src/checks/no_fixup.rs
@@ -18,7 +18,7 @@ pub struct NoFixup;
 
 use errors::*;
 use config::Checks;
-use github_handler::CommitData;
+use github::CommitData;
 use checks::Check;
 
 impl Check for NoFixup {
