+++ src/checks/no_squash.rs
@@ -18,7 +18,7 @@ pub struct NoSquash;
 
 use errors::*;
 use config::Checks;
-use github_handler::CommitData;
+use github::CommitData;
 use checks::Check;
 
 impl Check for NoSquash {
