+++ src/checks/no_wip.rs
@@ -18,7 +18,7 @@ pub struct NoWip;
 
 use errors::*;
 use config::Checks;
-use github_handler::CommitData;
+use github::CommitData;
 use checks::Check;
 
 impl Check for NoWip {
