+++ src/checks/requires_body.rs
@@ -16,7 +16,7 @@ pub struct RequiresBody;
 
 use errors::*;
 use config::Checks;
-use github_handler::CommitData;
+use github::CommitData;
 use checks::Check;
 
 impl Check for RequiresBody {
