+++ src/checks/summary_scope.rs
@@ -19,7 +19,7 @@ pub struct SummaryScope;
 use self::regex::Regex;
 use errors::*;
 use config::Checks;
-use github_handler::CommitData;
+use github::CommitData;
 use checks::Check;
 
 impl Check for SummaryScope {
